+++ src/types/generic/fragment.rs
@@ -1,9 +1,7 @@
 //! Fragment string.
 
-use crate::{
-    spec::Spec,
-    validate::{fragment, Error},
-};
+use crate::spec::Spec;
+use crate::validate::{fragment, Error, ErrorKind};
 
 define_custom_string_slice! {
     /// A borrowed slice of an IRI fragment (i.e. after the first `#` character).
@@ -40,10 +38,10 @@ define_custom_string_slice! {
     /// assert!(IriFragmentStr::new("#hash").is_err());
     /// ```
     ///
-    /// [RFC 3986]: https://tools.ietf.org/html/rfc3986
-    /// [RFC 3987]: https://tools.ietf.org/html/rfc3987
-    /// [`fragment` rule]: https://tools.ietf.org/html/rfc3986#section-3.5
-    /// [`ifragment` rule]: https://tools.ietf.org/html/rfc3987#section-2.2
+    /// [RFC 3986]: https://www.rfc-editor.org/rfc/rfc3986.html
+    /// [RFC 3987]: https://www.rfc-editor.org/rfc/rfc3987.html
+    /// [`fragment` rule]: https://www.rfc-editor.org/rfc/rfc3986.html#section-3.5
+    /// [`ifragment` rule]: https://www.rfc-editor.org/rfc/rfc3987.html#section-2.2
     struct RiFragmentStr {
         validator = fragment,
         expecting_msg = "IRI fragment string",
@@ -61,10 +59,10 @@ define_custom_string_owned! {
     ///
     /// Enabled by `alloc` or `std` feature.
     ///
-    /// [RFC 3986]: https://tools.ietf.org/html/rfc3986
-    /// [RFC 3987]: https://tools.ietf.org/html/rfc3987
-    /// [`fragment` rule]: https://tools.ietf.org/html/rfc3986#section-3.5
-    /// [`ifragment` rule]: https://tools.ietf.org/html/rfc3987#section-2.2
+    /// [RFC 3986]: https://www.rfc-editor.org/rfc/rfc3986.html
+    /// [RFC 3987]: https://www.rfc-editor.org/rfc/rfc3987.html
+    /// [`fragment` rule]: https://www.rfc-editor.org/rfc/rfc3986.html#section-3.5
+    /// [`ifragment` rule]: https://www.rfc-editor.org/rfc/rfc3987.html#section-2.2
     /// [`RiFragmentStr`]: struct.RiFragmentStr.html
     struct RiFragmentString {
         validator = fragment,
@@ -101,7 +99,7 @@ impl<S: Spec> RiFragmentStr<S> {
     /// ```
     pub fn from_prefixed(s: &str) -> Result<&Self, Error> {
         if !s.starts_with('#') {
-            return Err(Error::new());
+            return Err(Error::with_kind(ErrorKind::InvalidFragment));
         }
         TryFrom::try_from(&s[1..])
     }